+++ code/src/cljs/sixsq/nuvla/ui/main/events.cljs
@@ -17,6 +17,7 @@
 
 
 (def notification-polling-id :notifications-polling)
+(def check-ui-version-polling-id :check-ui-version)
 
 
 (reg-event-db
@@ -99,7 +100,9 @@
       {:db                        (assoc db ::spec/nav-path (split-path-alias path)
                                             ::spec/nav-query-params query-params)
        ::fx/bulk-actions-interval [::action-interval-delete
-                                   (dissoc actions-interval notification-polling-id)]})))
+                                   (dissoc actions-interval
+                                           notification-polling-id
+                                           check-ui-version-polling-id)]})))
 
 
 (reg-event-fx
@@ -275,29 +278,83 @@
 
 
 (reg-event-fx
-  ::get-ui-config-good
+  ::get-ui-config-success
   (fn [{db :db} [_ {:keys [stripe] :as result}]]
     (cond-> {:db (assoc db ::spec/config result)}
             stripe (assoc :dispatch [::load-stripe]))))
 
 
 (reg-event-db
-  ::get-ui-config-bad
-  (fn [db [_ _response]]
-    (log/info "Failed to load UI configuration file")
+  ::log-failed-response
+  (fn [db [_ msg response]]
+    (log/error msg ": " response)
     db))
+
+
+(reg-event-db
+  ::get-ui-version-success
+  (fn [{{:keys [current-version notify?]
+         :as   ui-version} ::spec/ui-version
+        :as                db} [_ response]]
+    (let [init-current-version? (nil? current-version)
+          new-version-detected? (and notify?
+                                     (some? current-version)
+                                     (not= current-version response))
+          should-open-modal?    (and notify?
+                                     new-version-detected?)]
+      (when init-current-version?
+        (log/info "Init ui version to: " response))
+      (when new-version-detected?
+        (log/info "New ui version detected: " response))
+      (assoc db ::spec/ui-version
+                (cond-> ui-version
+                        init-current-version? (assoc :current-version response)
+                        new-version-detected? (assoc :new-version response
+                                                     :notify? false)
+                        should-open-modal? (assoc :open-modal? true))))))
+
+
+(reg-event-db
+  ::new-version-open-modal?
+  (fn [db [_ open]]
+    (assoc-in db [::spec/ui-version :open-modal?] open)))
 
 
 (reg-event-fx
   ::get-ui-config
-  (fn [_ _]
-    {:http-xhrio {:method          :get
-                  :uri             "/ui/config/config.json"
-                  :timeout         8000
-                  :response-format (ajax/json-response-format {:keywords? true})
-                  :on-success      [::get-ui-config-good]
-                  :on-failure      [::get-ui-config-bad]}}))
+  (fn []
+    (let [force-no-cache (int (time/timestamp))
+          url            (str "/ui/config/config.json?" force-no-cache)]
+      {:http-xhrio {:method          :get
+                    :uri             url
+                    :timeout         8000
+                    :response-format (ajax/json-response-format {:keywords? true})
+                    :on-success      [::get-ui-config-success]
+                    :on-failure      [::log-failed-response
+                                      "Failed to load UI configuration file"]}})))
 
+
+(reg-event-fx
+  ::get-ui-version
+  (fn []
+    (let [force-no-cache (int (time/timestamp))
+          url            (str "/ui/css/version.css?" force-no-cache)]
+      {:http-xhrio {:method          :get
+                    :uri             url
+                    :timeout         8000
+                    :response-format (ajax/text-response-format)
+                    :on-success      [::get-ui-version-success]
+                    :on-failure      [::log-failed-response
+                                      "Failed to load UI version file"]}})))
+
+
+(reg-event-fx
+  ::check-ui-version-polling
+  (fn []
+    {:dispatch [::action-interval-start
+                {:id        check-ui-version-polling-id
+                 :frequency 120000
+                 :event     [::get-ui-version]}]}))
 
 (reg-event-db
   ::open-modal