+++ code/src/cljs/sixsq/nuvla/ui/main/views.cljs
@@ -203,6 +203,36 @@
        (panel/render @resource-path)])))
 
 
+(defn UpdateUIVersion
+  []
+  (let [tr          (subscribe [::i18n-subs/tr])
+        open?       (subscribe [::subs/ui-version-modal-open?])
+        new-version (subscribe [::subs/ui-version-new-version])
+        open-modal  #(dispatch [::events/new-version-open-modal? true])
+        close-modal #(dispatch [::events/new-version-open-modal? false])
+        reload      #(.reload js/location true)]
+    [:<>
+     [ui/Modal {:open @open?
+                :size :small
+                :basic true}
+      [uix/ModalHeader {:icon "refresh"
+                        :header (@tr [:new-ui-version])}]
+      [ui/ModalContent (@tr [:new-ui-version-content])]
+      [ui/ModalActions
+       [ui/Button {:negative true
+                   :on-click close-modal}
+        (@tr [:will-do-it-later])]
+       [ui/Button {:positive true
+                   :on-click reload}
+        (@tr [:update-and-reload])]]]
+     (when @new-version
+       [ui/MenuItem {:on-click open-modal}
+        [ui/Label {:color    "red"
+                   :icon     "refresh"
+                   :circular true
+                   :content  (@tr [:new-ui-version])}]])]))
+
+
 (defn header
   []
   [:header
@@ -216,7 +246,9 @@
 
     [ui/MenuItem [breadcrumbs]]
 
+
     [ui/MenuMenu {:position "right"}
+     [UpdateUIVersion]
      [messages/bell-menu]
      [session-views/AuthnMenu]]]